Add files via upload
This project is to build machine learning models on the byte and asm files to predict which type of malware these files represent. The byte files contain the hexadecimal codes and the asm file contains the assembly language code which contains keywords, opcodes, registers, APIs. We have to extract features from these files and build the optimal machine learning models on these extracted features.